Title: 자연수 n개의 합, Time: 84ms, Memory: 8MB, Status: Wrong Answer - CodeTree

+++ 240628/자연수 n개의 합/sum-of-n-natural-numbers.java
@@ -8,7 +8,7 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
 
-        s = Integer.parseInt(st.nextToken());
+        s = Long.parseLong(st.nextToken());
 
         long start = 1;
         long end = 200000000;
